error commit

+++ app/Http/Controllers/TweetController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
 use App\Http\Requests\TweetRequest;
+use App\Services;
+use App\Services\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class TweetController extends Controller
 {
@@ -20,9 +23,14 @@ class TweetController extends Controller
     public function create(TweetRequest $request, Tweet $tweet) {
 
         $tweet->content = $request->content;
+        $tweet->user_id = $request->userId();
         $tweet->save();
         return redirect()
             ->route('tweets.index');
+    }
+
+    public function add() {
+        return view('tweets.tweet');
     }
 
     public function show(Tweet $tweet) {
@@ -31,22 +39,31 @@ class TweetController extends Controller
             ->with('tweet', $tweet);
     }
 
-    public function edit(Tweet $tweet) {
+    public function edit(TweetRequest $request, Tweet $tweet, TweetService $tweetService) {
 
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweet->user_id)) {
+            throw new AccessDeniedHttpException();
+        }
         return view('tweets.edit')
             ->with('tweet', $tweet);
     }
 
-    public function update(TweetRequest $request, Tweet $tweet) {
+    public function update(TweetRequest $request, Tweet $tweet, TweetService $tweetService) {
 
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweet->user_id)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet->content = $request->content;
         $tweet->save();
         return redirect()
             ->route('tweets.show', $tweet);
     }
 
-    public function delete(Tweet $tweet) {
+    public function delete(TweetRequest $request, Tweet $tweet, TweetService $tweetService) {
 
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweet->user_id)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet->delete();
         return redirect()
             ->route('tweets.index');
